refactor(agent): keep composer draft under the file-length limit
Move session-sticky connector helpers and image conversion out of
agentComposerDraft.ts, and record the enable-set protocol in the
AgentGUI and runtimeprep architecture docs.
